Confirmacion borrar proyecto con cambios guardados

+++ practica_final/src/stores/projects.js
@@ -18,9 +18,34 @@ export const useProjectsStore = () => {
               })
           },
           async deleteProject (proyecto) {
-            const index = this.projects.indexOf(proyecto);
-            this.projects.splice(index, 1);
-            const responseProject = await axios.delete('http://localhost:8080/deleteProject?idProyecto=' + proyecto.id);
+            Swal.fire({
+              title: "¿Quieres borrar el proyecto?",
+              showDenyButton: true,
+              confirmButtonText: "Save",
+              denyButtonText: `Don't save`
+            }).then(async (result) => {
+              /* Read more about isConfirmed, isDenied below */
+              if (result.isConfirmed) {
+                try {
+                  const responseProject = await axios.delete('http://localhost:8080/deleteProject?idProyecto=' + proyecto.id);
+                  await this.getActiveProjects();      
+                  Swal.fire({
+                    title: 'Proyecto borrado',
+                    icon: 'success',
+                    draggable: true
+                  })
+                } catch (error) {
+                  this.dialog = false
+                  Swal.fire({
+                    title: error.response?.data?.error || 'Error al borrar proyecto',
+                    icon: 'error',
+                    draggable: true
+                  })
+                }
+              } else if (result.isDenied) {
+                Swal.fire("No se ha borrado el proyecto", "", "info");
+              }
+            });
           },
           async saveNewProject (proyecto) {
             let request = 'http://localhost:8080/createProject?desc=' + proyecto.descripcion + "&fInicio=" + proyecto.fInicio
